requirements.txt
flask
langgraph
langchain-groq
yfinance
pandas
python-dotenv
statsmodels
numpy-financial
requests
